requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
csvkit==1.1.1
defusedxml==0.7.1
Django==3.2.18
django-cors-headers==3.14.0
django-filter==2.4.0
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
firebase_admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google_cloud_firestore==2.11.1
google_cloud_storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.21.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==1.7.0
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
mccabe==0.7.0
msgpack==1.0.5
oauthlib==3.2.2
onesignal-python-api==2.0.0
parsedatetime==2.6
phonenumberslite==8.13.6
Pillow==9.4.0
protobuf==4.22.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyfcm==1.5.4
pyflakes==2.5.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.1
python-telegram-bot==20.2
python3-openid==3.2.0
pytimeparse==1.1.8
pytz==2022.7.1
pyyaml==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.4.0